apps/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
crewai==0.22.5
crewai-tools==0.4.26
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
nats-py==2.6.0
boto3==1.34.0
clickhouse-driver==0.2.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
sentry-sdk[fastapi]==1.38.0
